extractor/requirements.txt
torch
pdf2image
redis
flask
opencv2-python
